+++ lib/mutant/mutator/node/when.rb
@@ -16,7 +16,32 @@ module Mutant
         # @api private
         #
         def dispatch
-          emit_attribute_mutations(:body)
+          mutate_body
+          mutate_conditions
+        end
+
+        # Emit condition mutations
+        #
+        # @return [undefined]
+        #
+        # @api private
+        #
+        def mutate_conditions
+          conditions = children.length - 1
+          children[0..-2].each_index do |index|
+            delete_child(index) if conditions > 1
+            mutate_child(index)
+          end
+        end
+
+        # Emit body mutations
+        #
+        # @return [undefined]
+        #
+        # @api private
+        #
+        def mutate_body
+          mutate_child(children.length-1)
         end
 
       end # When